Moved useNavPrompt to hooks
