+++ lib/components/formik/BlockNavigation/index.js
@@ -2,7 +2,7 @@ import React from "react";
 import { useFormikContext } from "formik";
 import PropTypes from "prop-types";
 
-import useNavPrompt from "./useNavPrompt";
+import useNavPrompt from "../../../hooks/useNavPrompt";
 import Alert from "../../Alert";
 
 const BlockNavigation = ({ isDirty = false }) => {
